Complete HotSwap

+++ nginx/swap.sh
@@ -2,10 +2,11 @@
 
 cd /etc/nginx
 
-if [ -f ./nginx.conf.bak ]; then
+if [ -f ./nginx.conf.bak && -z $1 ]; then
     mv nginx.conf.bak nginx.conf
-else
-    sed -i.bak "s|proxy_pass http://activity:8080/activity/|proxy_pass http://activity2:8080/activity/|" nginx.conf
+elif [ -z $1 ]; then
+    sed -i.bak "s|proxy_pass.*|proxy_pass http://${1}:8080/activity/|" nginx.conf
 fi
 
+#TODO: check if the service faile to run after hot swap
 service nginx reload
